test: refactor common functions for env and addons
